Criei duas novas class que herdam da class User
