+++ pagina-login/cadastro/src/main/java/com/example/cadastro/model/Candidate.java
@@ -1,0 +1,36 @@
+package com.example.cadastro.model;
+import java.time.LocalDate;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Past;
+import jakarta.validation.constraints.Size;
+
+@Entity
+public class Candidate extends User {
+    @Id
+    @NotBlank(message = "CPF é obrigatório")
+    @Size(min = 11, max = 11, message = "CPF deve ter 11 caracteres")
+    private String cpf;
+
+    @Past(message = "Data de nascimento deve ser no passado")
+    private LocalDate dataNacimento;
+
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public LocalDate getDataNacimento() {
+        return dataNacimento;
+    }
+
+    public void setDataNacimento(LocalDate dataNacimento) {
+        this.dataNacimento = dataNacimento;
+    }
+    
+}
