+++ pagina-login/cadastro/src/main/java/com/example/cadastro/model/Contractor.java
@@ -1,0 +1,33 @@
+package com.example.cadastro.model;
+
+public class Contractor extends User {
+    private String cnpj;
+    private String empresa;
+    private String telefone;
+
+    public String getCnpj() {
+        return cnpj;
+    }
+
+    public void setCnpj(String cnpj) {
+        this.cnpj = cnpj;
+    }
+
+    public String getEmpresa() {
+        return empresa;
+    }
+
+    public void setEmpresa(String empresa) {
+        this.empresa = empresa;
+    }
+
+    public String getTelefone() {
+        return telefone;
+    }
+
+    public void setTelefone(String telefone) {
+        this.telefone = telefone;
+    }
+
+    
+}
