+++ pagina-login/cadastro/src/main/java/com/example/cadastro/security/JwtUtil.java
@@ -1,15 +1,9 @@
 package com.example.cadastro.security;
-
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
-
-import org.hibernate.boot.model.source.spi.SingularAttributeNature;
 import org.springframework.stereotype.Component;
-
-import com.fasterxml.jackson.databind.cfg.ConstructorDetector.SingleArgConstructor;
-//fgdgsfda
 import java.security.Key;
 import java.util.Date;
